fix: refactored code to fix auto imports
