+++ tsconfig.json
@@ -7,6 +7,7 @@
     "emitDeclarationOnly": false,
     "moduleResolution": "node",
     "esModuleInterop": true,
+    "allowSyntheticDefaultImports": true,
     "module": "ESNext",
     "target": "ESNext"
   },
